feat: add database routes and admin API for user and API key management
- Implemented new database routes in `src/database/routes.ts` for managing users and API keys.
- Added CRUD operations for users and API keys, including validation and error handling.
- Introduced new schemas for user and API key creation and updates using Zod.
- Updated environment configuration to include `D1_DATABASE` and logging options.
- Created a README file for the database API with example curl commands for common operations.
