+++ backend/package.json
@@ -32,7 +32,8 @@
     "db:init": "npx tsx scripts/init-database.ts",
     "db:seed": "wrangler d1 execute d1-dex-database --local --file=./scripts/init-db.sql",
     "db:query": "wrangler d1 execute d1-dex-database --local",
-    "db:query:prod": "wrangler d1 execute d1-dex-database",
+    "db:migration:prod": "wrangler d1 execute d1-dex-database --remote --file drizzle/0000_stale_infant_terrible.sql",
+    "db:query:prod": "wrangler d1 execute d1-dex-database --command \"SELECT name FROM sqlite_master WHERE type='table';\"",
     "db:backup": "wrangler d1 export d1-dex-database --output=backup.sql",
     "db:reset": "wrangler d1 execute d1-dex-database --local --command='DROP SCHEMA IF EXISTS main CASCADE; CREATE SCHEMA main;'",
     "verify": "./scripts/verify-config.sh",
