+++ backend/worker-configuration.d.ts
@@ -1,11 +1,16 @@
 /* eslint-disable */
-// Generated by Wrangler by running `wrangler types` (hash: 832f67d57ef6a0470d669fe20e0e5a0b)
+// Generated by Wrangler by running `wrangler types` (hash: c8a1eba694f7517d3ada041e2f66c998)
 // Runtime types generated with workerd@1.20250617.0 2025-06-20 nodejs_compat
 declare namespace Cloudflare {
 	interface Env {
-		NODE_ENV: "production" | "development" | "staging";
+		NODE_ENV: "production" | "development";
+		LOG_LEVEL: "info" | "debug" | "warn";
+		ENABLE_CONSOLE_LOGS: "true" | "false";
+		ENABLE_STRUCTURED_LOGS: "true";
 		SUBGRAPH_URL: "https://api.studio.thegraph.com/query/114739/entysquare-dex-bsc-testnet/version/latest";
 		SUBGRAPH_HEALTH_URL: "https://indexer.upgrade.thegraph.com/status";
+		R2: R2Bucket;
+		D1_DATABASE: D1Database;
 		AI: Ai;
 	}
 }
@@ -14,7 +19,7 @@ type StringifyValues<EnvType extends Record<string, unknown>> = {
 	[Binding in keyof EnvType]: EnvType[Binding] extends string ? EnvType[Binding] : string;
 };
 declare namespace NodeJS {
-	interface ProcessEnv extends StringifyValues<Pick<Cloudflare.Env, "NODE_ENV" | "SUBGRAPH_URL" | "SUBGRAPH_HEALTH_URL">> {}
+	interface ProcessEnv extends StringifyValues<Pick<Cloudflare.Env, "NODE_ENV" | "LOG_LEVEL" | "ENABLE_CONSOLE_LOGS" | "ENABLE_STRUCTURED_LOGS" | "SUBGRAPH_URL" | "SUBGRAPH_HEALTH_URL">> {}
 }
 
 // Begin runtime types
